:recycle: ref: pull slack thread util methods into a seperate class
